// src/app/payment-module/model/payment-method-model.ts
export class PaymentMethodModel {

  dateCreated: any;
  dateUpdated: any;
  enabled: boolean;
  id: number;
  live: boolean;
  liveActive: boolean;
  livePublicKey: string;
  liveSecret: string;
  liveVerifyUrl: string;
  paymentMethodName: string;
  testingPublicKey: string;
  testingSecret: string;
  testingVerifyUrl: string;
}